feat: added user toxicity check
vkapi/vk-tools.py: added a method for receiving
user posts and comments and checking them for
the presence of obscene vocabulary.
toxicity_check.py: the function of checking the
text for the presence of obscene vocabulary from
a remote database has been added.
The file 'requirements.txt' has been updated.