feat: added acquaintances searching with gigachat
Added the 'analyse_acquaintances' method in the Vk class,
which selects from an array of VK accounts those that,
according to gigachat, will be interesting to this user for
acquaintance.
Added the 'check_acquaintances' function to connect to the
giga chat api and send a request to verify user interests
in a file 'gigachat_tools.py'.
Added a JSON database of a large number of VK users for
quick dating analysis.
Added a method for generating this database in the Vk class
('__dump_big_users_data').